+++ esphome/components/xiaomi_xmtzc0xhm/__init__.py
