+++ esphome/components/xiaomi_xmtzc0xhm/sensor.py
@@ -10,8 +10,8 @@ from esphome.const import CONF_BATTERY_LEVEL, CONF_MAC_ADDRESS, \
 DEPENDENCIES = ['esp32_ble_tracker']
 AUTO_LOAD = ['xiaomi_ble']
 
-xiaomi_miscale_ns = cg.esphome_ns.namespace('xiaomi_miscale')
-XiaomiMiscale = xiaomi_miscale_ns.class_('XiaomiMiscale', esp32_ble_tracker.ESPBTDeviceListener,
+xiaomi_xmtzc0xhm_ns = cg.esphome_ns.namespace('xiaomi_xmtzc0xhm')
+XiaomiMiscale = xiaomi_xmtzc0xhm_ns.class_('XiaomiMiscale', esp32_ble_tracker.ESPBTDeviceListener,
                                          cg.Component)
 
 CONFIG_SCHEMA = cv.Schema({
