+++ esphome/components/xiaomi_xmtzc0xhm/xiaomi_xmtzc0xhm.cpp
@@ -1,19 +1,19 @@
-#include "xiaomi_miscale.h"
+#include "xiaomi_xmtzc0xhm.h"
 #include "esphome/core/log.h"
 
 #ifdef ARDUINO_ARCH_ESP32
 
 namespace esphome {
-namespace xiaomi_miscale {
+namespace xiaomi_xmtzc0xhm {
 
-static const char *TAG = "xiaomi_miscale";
+static const char *TAG = "xiaomi_xmtzc0xhm";
 
 void XiaomiMiscale::dump_config() {
   ESP_LOGCONFIG(TAG, "Xiaomi Mijia");
   LOG_SENSOR("  ", "Measured Weight", this->weight_);
 }
 
-}  // namespace xiaomi_miscale
+}  // namespace xiaomi_xmtzc0xhm
 }  // namespace esphome
 
 #endif
